+++ haas/haas_application.py
@@ -97,10 +97,8 @@ class HaasApplication(object):
 
         args = self.parser.parse_args(self.argv[1:])
 
-        plugin_manager.configure_plugins(args)
-
         environment_plugins = plugin_manager.get_enabled_hook_plugins(
-            plugin_manager.ENVIRONMENT_HOOK)
+            plugin_manager.ENVIRONMENT_HOOK, args)
         runner = plugin_manager.get_driver(
             plugin_manager.TEST_RUNNER, args)
 
@@ -121,12 +119,14 @@ class HaasApplication(object):
             else:
                 suite = loader.create_suite(suites)
             test_count = suite.countTestCases()
-            result_handler = plugin_manager.get_driver(
+            result_handlers = plugin_manager.get_enabled_hook_plugins(
                 plugin_manager.RESULT_HANDLERS, args, test_count=test_count)
 
             result_collector = ResultCollecter(
                 buffer=args.buffer, failfast=args.failfast)
-            result_collector.add_result_handler(result_handler)
+
+            for result_handler in result_handlers:
+                result_collector.add_result_handler(result_handler)
 
             result = runner.run(result_collector, suite)
             return not result.wasSuccessful()